Split javax.servlet/HttpService support out of core
Core no longer contains or directly selects the classic javax.servlet or
OSGi HttpService implementation -- previously that was only made optional
at the OSGi manifest level, with the actual classes (ServletStreamServerImpl,
HttpServiceServletContainerAdapter, etc.) still compiled into and shipped
with org.jupnp. Two new optional bundles take over:

- org.jupnp.transport.servlet: the generic javax.servlet transport
  infrastructure (ServletStreamServerImpl, the ServletContainerAdapter SPI,
  the blocking/async servlet variants). Genuinely shared, not Jetty-specific:
  both org.jupnp.transport.jetty9 and org.jupnp.android depend on it.
- org.jupnp.transport.httpservice: the classic OSGi HttpService bridge
  (HttpServiceServletContainerAdapter, DisableAuthenticationHttpContext),
  plus a new HttpServiceSharedStreamServerProvider Declarative Services
  component that only activates while a real HttpService is bound.

Core gets a new generic org.jupnp.transport.spi.SharedStreamServerProvider
SPI instead: OSGiUpnpServiceConfiguration.createStreamServer() consults it
(same ambiguity-detection pattern as the existing TransportConfiguration
reference) and falls through to the discovered transport's own native
StreamServer when none is bound, with no compile-time or runtime dependency
on javax.servlet or HttpService of any kind. org.jupnp.http.RequestInfo was
split the same way: the String-based device-detection helpers
RemoteClientInfo actually uses stayed in core; the HttpServletRequest-based
counterparts moved to the new servlet bundle as ServletRequestInfo.

This also fully resolves a gap the previous "optional import" approach had:
Declarative Services runtimes (e.g. Apache Felix SCR) locate bind methods
via Class.getDeclaredMethods(), which must resolve every declared method's
parameter/return types -- including ones for references that are never
satisfied. A method literally typed HttpService threw NoClassDefFoundError
out of that reflective call on a genuinely HttpService-less runtime, taking
down introspection of the whole OSGiUpnpServiceConfiguration component, not
just the optional reference. Moving the HttpService-typed code into its own
bundle with ordinary mandatory imports sidesteps this entirely: on a runtime
without HttpService, that bundle simply fails to resolve and none of its
classes are ever introspected.

Also fixes three split-package bugs the move initially introduced and that
running the real OSGi integration tests caught: org.jupnp.transport.spi,
org.jupnp.transport.impl, and org.jupnp.http were each briefly shared
between core and a new bundle, which OSGi doesn't support (an importer
resolves to exactly one provider per package). The servlet bundle's classes
now live under its own org.jupnp.transport.impl.servlet namespace instead.

Updates jetty9's and Android's poms with the new dependency, and fixes
website/docs/advanced.md's code examples and two stale core javadoc
comments (still describing a "default transport implementation found in
org.jupnp.transport.impl" from before transports were modularized at all).

Verified with a full clean reactor build (checkstyle, spotless, license all
enabled) and by running both OSGi integration tests for real (actual
Equinox + Felix SCR + Aries SPI Fly), not just compiling.
