+++ bundles/org.jupnp.android/src/main/java/org/jupnp/android/AndroidUpnpServiceConfiguration.java
@@ -24,11 +24,11 @@ import org.jupnp.model.Namespace;
 import org.jupnp.model.ServerClientTokens;
 import org.jupnp.transport.impl.GENAEventProcessorImpl;
 import org.jupnp.transport.impl.SOAPActionProcessorImpl;
-import org.jupnp.transport.impl.ServletStreamServerConfigurationImpl;
-import org.jupnp.transport.impl.ServletStreamServerImpl;
 import org.jupnp.transport.impl.jetty.JettyServletContainer;
 import org.jupnp.transport.impl.jetty.JettyStreamClientImpl;
 import org.jupnp.transport.impl.jetty.StreamClientConfigurationImpl;
+import org.jupnp.transport.impl.servlet.ServletStreamServerConfigurationImpl;
+import org.jupnp.transport.impl.servlet.ServletStreamServerImpl;
 import org.jupnp.transport.spi.GENAEventProcessor;
 import org.jupnp.transport.spi.NetworkAddressFactory;
 import org.jupnp.transport.spi.SOAPActionProcessor;
