+++ bundles/org.jupnp.transport.httpservice/bnd.bnd
@@ -1,0 +1,1 @@
+-include: ../../bnd.bnd
