+++ bundles/org.jupnp.transport.httpservice/pom.xml
@@ -10,9 +10,9 @@
   </parent>
 
   <groupId>org.jupnp</groupId>
-  <artifactId>org.jupnp.transport.jetty9</artifactId>
+  <artifactId>org.jupnp.transport.httpservice</artifactId>
 
-  <name>jUPnP Jetty 9.4.x Transport</name>
+  <name>jUPnP OSGi HttpService Bridge</name>
 
   <properties>
     <basedirRoot>../..</basedirRoot>
@@ -25,8 +25,10 @@
       <version>${project.version}</version>
       <scope>compile</scope>
     </dependency>
-    <!-- For ServletStreamServerImpl/ServletContainerAdapter: createStreamServer() registers a Jetty-backed
-         adapter with the generic servlet transport infrastructure, which core no longer bundles. -->
+    <!--
+      For ServletStreamServerImpl/ServletContainerAdapter/AsyncServlet: the classic OSGi HttpService only
+      registers javax.servlet Servlets, so bridging to it needs the generic servlet transport support too.
+    -->
     <dependency>
       <groupId>org.jupnp</groupId>
       <artifactId>org.jupnp.transport.servlet</artifactId>
@@ -34,35 +36,10 @@
       <scope>compile</scope>
     </dependency>
     <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-client</artifactId>
-      <version>${jetty.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <version>${jetty.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-servlet</artifactId>
-      <version>${jetty.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
       <version>${slf4j.version}</version>
       <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jupnp</groupId>
-      <artifactId>org.jupnp</artifactId>
-      <version>${project.version}</version>
-      <type>test-jar</type>
-      <scope>test</scope>
     </dependency>
   </dependencies>
