+++ bundles/org.jupnp.transport.httpservice/src/main/java/org/jupnp/transport/impl/osgi/DisableAuthenticationHttpContext.java
