+++ bundles/org.jupnp.transport.jetty9/src/main/java/org/jupnp/transport/impl/jetty/JettyServletContainer.java
@@ -24,7 +24,7 @@ import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
-import org.jupnp.transport.spi.ServletContainerAdapter;
+import org.jupnp.transport.impl.servlet.ServletContainerAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
