+++ bundles/org.jupnp.transport.servlet/bnd.bnd
@@ -1,0 +1,1 @@
+-include: ../../bnd.bnd
