+++ bundles/pom.xml
@@ -20,6 +20,8 @@
     <module>org.jupnp.android</module>
     <module>org.jupnp.osgi</module>
     <module>org.jupnp.support</module>
+    <module>org.jupnp.transport.servlet</module>
+    <module>org.jupnp.transport.httpservice</module>
     <module>org.jupnp.transport.jetty9</module>
   </modules>
